+++ CHANGES.rst
@@ -3,10 +3,26 @@ History
 =======
 
 
-1.4.2 (unreleased)
-------------------
+2.0 (unreleased)
+----------------
 
-- Nothing changed yet.
+.. warning:: Upgrade Warning!
+    After installing this release, you need to run the provided upgrade steps.
+
+New:
+
+- Set the childsite skin on a ``lineage.registry`` based local registry instead of an ``lineage_theme`` attribute on an ``IChildSite`` object.
+  Includes an upgrade step.
+  [thet]
+
+- Allow setting of ``lineage_theme`` skin setting through the ``@@theming-controlpanel`` and remove the customized controlpanel template, Archetypes schemaextender and Dexterity behavior.
+  This works for ``plone.app.theming`` ``1.1.x`` and ``1.2.x``.
+  Includes an upgrade step.
+  [thet]
+
+Fixes:
+
+- *add item here*
 
 
 1.4.1 (2016-02-08)